Keep compiled Flex create_block_mask and SM90 FA4 dispatch. Drop 1142's eager-only Flex guards that belong on the registry PR, not this branch.

+++ veomni/models/transformers/deepseek_v3/deepseek_v3_gpu_patch_gen_config.py
@@ -250,7 +250,7 @@ class PatchedDeepseekV3Experts(nn.Module):
             return merged_experts_act_fn_forward(
                 hidden_states,
                 top_k_index,
-                top_k_weights,
+                top_k_weights.to(hidden_states.dtype),
                 self.gate_up_proj,
                 self.down_proj,
                 self.act_fn,
@@ -259,7 +259,7 @@ class PatchedDeepseekV3Experts(nn.Module):
         unused = self.gate_up_proj.new_empty(0)
         return self.veomni_moe(
             hidden_states,
-            top_k_weights,
+            top_k_weights.to(hidden_states.dtype),
             top_k_index,
             unused,
             unused,
